Fixed multiple bugs including resizing crash, crash when window size is 0, 0 and also cleaned up a bit of code related to GPU querying. #9 can be marked as solved since development is done on a single machine and therefore, I don't have access to hardware with multiple GPUs. Once the time comes, then this particular issue can be looked at.

+++ vmve/src/ui.hpp
@@ -13,8 +13,8 @@
 
 extern bool viewportActive;
 extern bool resizeViewport;
-extern float viewport_width;
-extern float viewport_height;
+extern int viewport_width;
+extern int viewport_height;
